Generic group logs parser
